Fix registration payment bug and add loading spinner
This commit fixes a bug in the `updateRegistrationPaymentId` function where the `items` object was not being properly updated. It now correctly updates the `code` property of the first item in the `items` array.

In the `EventPage` component, the registration button is now disabled when there is a registration error of type "error". Additionally, a loading spinner is displayed when generating a new payment.

In the `PaymentModal` component, a loading spinner is displayed when generating a new payment.